Task/move market data endpoint to asset profiles (#7026 follow-up)
