+++ apps/api/src/app/endpoints/asset-profiles/asset-profiles.module.ts
@@ -1,7 +1,11 @@
 import { ActivitiesModule } from '@ghostfolio/api/app/activities/activities.module';
+import { TransformDataSourceInRequestModule } from '@ghostfolio/api/interceptors/transform-data-source-in-request/transform-data-source-in-request.module';
+import { TransformDataSourceInResponseModule } from '@ghostfolio/api/interceptors/transform-data-source-in-response/transform-data-source-in-response.module';
 import { ApiModule } from '@ghostfolio/api/services/api/api.module';
 import { BenchmarkModule } from '@ghostfolio/api/services/benchmark/benchmark.module';
+import { DataProviderModule } from '@ghostfolio/api/services/data-provider/data-provider.module';
 import { ExchangeRateDataModule } from '@ghostfolio/api/services/exchange-rate-data/exchange-rate-data.module';
+import { MarketDataModule } from '@ghostfolio/api/services/market-data/market-data.module';
 import { PrismaModule } from '@ghostfolio/api/services/prisma/prisma.module';
 import { SymbolProfileModule } from '@ghostfolio/api/services/symbol-profile/symbol-profile.module';
 
@@ -16,10 +20,15 @@ import { AssetProfilesService } from './asset-profiles.service';
     ActivitiesModule,
     ApiModule,
     BenchmarkModule,
+    DataProviderModule,
     ExchangeRateDataModule,
+    MarketDataModule,
     PrismaModule,
-    SymbolProfileModule
+    SymbolProfileModule,
+    TransformDataSourceInRequestModule,
+    TransformDataSourceInResponseModule
   ],
-  providers: [AssetProfilesService]
+  providers: [AssetProfilesService],
+  exports: [AssetProfilesService]
 })
 export class AssetProfilesModule {}
