+++ apps/api/src/app/endpoints/market-data/market-data.module.ts
@@ -1,7 +1,4 @@
-import { AdminModule } from '@ghostfolio/api/app/admin/admin.module';
 import { SymbolModule } from '@ghostfolio/api/app/symbol/symbol.module';
-import { TransformDataSourceInRequestModule } from '@ghostfolio/api/interceptors/transform-data-source-in-request/transform-data-source-in-request.module';
-import { TransformDataSourceInResponseModule } from '@ghostfolio/api/interceptors/transform-data-source-in-response/transform-data-source-in-response.module';
 import { MarketDataModule as MarketDataServiceModule } from '@ghostfolio/api/services/market-data/market-data.module';
 import { SymbolProfileModule } from '@ghostfolio/api/services/symbol-profile/symbol-profile.module';
 
@@ -11,13 +8,6 @@ import { MarketDataController } from './market-data.controller';
 
 @Module({
   controllers: [MarketDataController],
-  imports: [
-    AdminModule,
-    MarketDataServiceModule,
-    SymbolModule,
-    SymbolProfileModule,
-    TransformDataSourceInRequestModule,
-    TransformDataSourceInResponseModule
-  ]
+  imports: [MarketDataServiceModule, SymbolModule, SymbolProfileModule]
 })
 export class MarketDataModule {}
